+++ src/Settings/DataProcessingSettingsView.tsx
@@ -1,23 +1,29 @@
-import React from 'react';
+import React from "react";
 import {
   Button,
   Checkbox,
   FormControl,
   FormControlLabel,
+  FormLabel,
   InputAdornment,
   InputLabel,
   MenuItem,
+  Radio,
+  RadioGroup,
   Select,
   TextField,
   Tooltip,
-} from '@mui/material';
-import { observer } from 'mobx-react-lite';
+} from "@mui/material";
+import { observer } from "mobx-react-lite";
 
-import { App } from 'src/App';
+import { App } from "src/App";
 import {
+  CarriageReturnCursorBehaviors,
   DataViewConfiguration,
+  NewLineCursorBehaviors,
   dataViewConfigEnumToDisplayName,
-} from 'src/Settings/DataProcessingSettings';
+} from "src/Settings/DataProcessingSettings";
+import BorderedSection from "src/Components/BorderedSection";
 
 interface Props {
   app: App;
@@ -27,7 +33,9 @@ function DataProcessingView(props: Props) {
   const { app } = props;
 
   return (
-    <div style={{ display: 'flex', flexDirection: 'column', alignItems: 'start' }}>
+    <div
+      style={{ display: "flex", flexDirection: "column", alignItems: "start" }}
+    >
       {/* =============================================================================== */}
       {/* ANSI ESCAPE CODE PARSING ENABLED */}
       {/* =============================================================================== */}
@@ -54,12 +62,14 @@ function DataProcessingView(props: Props) {
             />
           }
           label="ANSI Escape Code Parsing"
-          sx={{ marginBottom: '10px' }}/>
+          sx={{ marginBottom: "10px" }}
+        />
       </Tooltip>
       {/* =============================================================================== */}
       {/* MAX. ESCAPE CODE LENGTH */}
       {/* =============================================================================== */}
-      <Tooltip title="The max. length of escape code allowed (in characters). Certain malformed escape codes (or data interruptions) could cause the escape code parser to get stuck thinking the incoming data stream is part of an escape code. This limit is so that at a certain length the parser rejects the partial code and goes back to the IDLE state. This includes all characters in the escape code, including the starting \x1B byte. Must be a least 2 chars."
+      <Tooltip
+        title="The max. length of escape code allowed (in characters). Certain malformed escape codes (or data interruptions) could cause the escape code parser to get stuck thinking the incoming data stream is part of an escape code. This limit is so that at a certain length the parser rejects the partial code and goes back to the IDLE state. This includes all characters in the escape code, including the starting \x1B byte. Must be a least 2 chars."
         followCursor
         arrow
       >
@@ -92,13 +102,14 @@ function DataProcessingView(props: Props) {
             app.settings.dataProcessing.visibleData.fields
               .maxEscapeCodeLengthChars.errorMsg
           }
-          sx={{ marginBottom: '20px' }}
+          sx={{ marginBottom: "20px" }}
         />
       </Tooltip>
       {/* =============================================================================== */}
       {/* DATA WIDTH */}
       {/* =============================================================================== */}
-      <Tooltip title="The max. number of characters to display per line in the terminal before wrapping to the next line. Must be a positive integer. New line characters also cause text to jump to the next line."
+      <Tooltip
+        title="The max. number of characters to display per line in the terminal before wrapping to the next line. Must be a positive integer. New line characters also cause text to jump to the next line."
         followCursor
         arrow
       >
@@ -114,8 +125,8 @@ function DataProcessingView(props: Props) {
             ),
           }}
           value={
-            app.settings.dataProcessing.visibleData.fields
-              .terminalWidthChars.value
+            app.settings.dataProcessing.visibleData.fields.terminalWidthChars
+              .value
           }
           onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
             app.settings.dataProcessing.onFieldChange(
@@ -124,20 +135,21 @@ function DataProcessingView(props: Props) {
             );
           }}
           error={
-            app.settings.dataProcessing.visibleData.fields
-              .terminalWidthChars.hasError
+            app.settings.dataProcessing.visibleData.fields.terminalWidthChars
+              .hasError
           }
           helperText={
-            app.settings.dataProcessing.visibleData.fields
-              .terminalWidthChars.errorMsg
+            app.settings.dataProcessing.visibleData.fields.terminalWidthChars
+              .errorMsg
           }
-          sx={{ marginBottom: '20px' }}
+          sx={{ marginBottom: "20px" }}
         />
       </Tooltip>
       {/* =============================================================================== */}
       {/* CHAR SIZE */}
       {/* =============================================================================== */}
-      <Tooltip title="The font size (in pixels) of characters displayed in the terminal."
+      <Tooltip
+        title="The font size (in pixels) of characters displayed in the terminal."
         followCursor
         arrow
       >
@@ -148,13 +160,9 @@ function DataProcessingView(props: Props) {
           variant="outlined"
           size="small"
           InputProps={{
-            endAdornment: (
-              <InputAdornment position="start">px</InputAdornment>
-            ),
+            endAdornment: <InputAdornment position="start">px</InputAdornment>,
           }}
-          value={
-            app.settings.dataProcessing.charSizePx.dispValue
-          }
+          value={app.settings.dataProcessing.charSizePx.dispValue}
           onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
             app.settings.dataProcessing.setCharSizePxDisp(event.target.value);
           }}
@@ -166,19 +174,16 @@ function DataProcessingView(props: Props) {
               app.settings.dataProcessing.applyCharSizePx();
             }
           }}
-          error={
-            app.settings.dataProcessing.charSizePx.hasError
-          }
-          helperText={
-            app.settings.dataProcessing.charSizePx.errorMsg
-          }
-          sx={{ marginBottom: '20px' }}
+          error={app.settings.dataProcessing.charSizePx.hasError}
+          helperText={app.settings.dataProcessing.charSizePx.errorMsg}
+          sx={{ marginBottom: "20px" }}
         />
       </Tooltip>
       {/* =============================================================================== */}
       {/* SCROLLBACK BUFFER SIZE */}
       {/* =============================================================================== */}
-      <Tooltip title="The max. number of rows to store in any terminal scrollback buffer (TX, RX, TX/RX).
+      <Tooltip
+        title="The max. number of rows to store in any terminal scrollback buffer (TX, RX, TX/RX).
         Increasing this will give you more history but decrease performance and increase memory usage. Must be a positive non-zero integer."
         followCursor
         arrow
@@ -211,7 +216,7 @@ function DataProcessingView(props: Props) {
             app.settings.dataProcessing.visibleData.fields
               .scrollbackBufferSizeRows.errorMsg
           }
-          sx={{ marginBottom: '20px' }}
+          sx={{ marginBottom: "20px" }}
         />
       </Tooltip>
       {/* =============================================================================== */}
@@ -223,7 +228,7 @@ function DataProcessingView(props: Props) {
         followCursor
         arrow
       >
-        <FormControl size="small" sx={{ minWidth: '210px' }}>
+        <FormControl size="small" sx={{ minWidth: "210px" }}>
           <InputLabel>Data View Configuration</InputLabel>
           <Select
             name="dataViewConfiguration"
@@ -237,7 +242,7 @@ function DataProcessingView(props: Props) {
                 Number(e.target.value)
               );
             }}
-            sx={{ marginBottom: '20px' }}
+            sx={{ marginBottom: "20px" }}
           >
             {Object.keys(DataViewConfiguration)
               .filter((key) => !Number.isNaN(Number(key)))
@@ -262,24 +267,24 @@ function DataProcessingView(props: Props) {
         followCursor
         arrow
       >
-      <FormControlLabel
-        control={
-          <Checkbox
-            name="localTxEcho"
-            checked={
-              app.settings.dataProcessing.visibleData.fields
-                .localTxEcho.value
-            }
-            onChange={(e) => {
-              app.settings.dataProcessing.onFieldChange(
-                e.target.name,
-                e.target.checked
-              );
-            }}
-          />
-        }
-        label="Local TX Echo"
-        sx={{ marginBottom: '10px' }}/>
+        <FormControlLabel
+          control={
+            <Checkbox
+              name="localTxEcho"
+              checked={
+                app.settings.dataProcessing.visibleData.fields.localTxEcho.value
+              }
+              onChange={(e) => {
+                app.settings.dataProcessing.onFieldChange(
+                  e.target.name,
+                  e.target.checked
+                );
+              }}
+            />
+          }
+          label="Local TX Echo"
+          sx={{ marginBottom: "10px" }}
+        />
       </Tooltip>
       {/* =============================================================================== */}
       {/* APPLY BUTTON */}
@@ -294,6 +299,184 @@ function DataProcessingView(props: Props) {
       >
         Apply
       </Button>
+
+      {/* Row with new line and carriage return settings */}
+      <div style={{ display: 'flex' }}>
+      {/* =============================================================================== */}
+      {/* NEW LINE SETTINGS */}
+      {/* =============================================================================== */}
+      <BorderedSection title="New Lines">
+        <div
+          style={{
+            display: "flex",
+            flexDirection: "column",
+            maxWidth: "300px",
+            gap: "20px",
+          }}
+        >
+          {/* NEW LINE BEHAVIOR */}
+          <FormControl>
+            <FormLabel>When a \n byte is received:</FormLabel>
+            <RadioGroup
+              value={app.settings.dataProcessing.newLineCursorBehavior}
+              onChange={(e) => {
+                app.settings.dataProcessing.setNewLineBehavior(
+                  e.target.value as any
+                );
+              }}
+            >
+              {/* DO NOTHING */}
+              <Tooltip
+                title="Don't move the cursor at all when a new line character is received."
+                placement="right"
+                arrow
+              >
+                <FormControlLabel
+                  value={NewLineCursorBehaviors.DO_NOTHING}
+                  control={<Radio />}
+                  label="Don't move the cursor"
+                  data-testid="new-line-dont-move-cursor"
+                />
+              </Tooltip>
+              {/* MOVE DOWN ONE LINE */}
+              <Tooltip
+                title="Move the cursor directly down one line. A separate carriage return is required if you want to move the cursor to the start of the new line."
+                placement="right"
+                arrow
+              >
+                <FormControlLabel
+                  value={NewLineCursorBehaviors.NEW_LINE}
+                  control={<Radio />}
+                  label="Move cursor down one line (new line)"
+                />
+              </Tooltip>
+              {/* NEW LINE AND CARRIAGE RETURN */}
+              <Tooltip
+                title="Move the cursor back to the start of the line and then down one line. This is the most common behavior for receiving a new line character."
+                placement="right"
+                arrow
+              >
+                <FormControlLabel
+                  value={NewLineCursorBehaviors.CARRIAGE_RETURN_AND_NEW_LINE}
+                  control={<Radio />}
+                  label="Move cursor to the start of line and then down one line."
+                />
+              </Tooltip>
+            </RadioGroup>
+          </FormControl>
+          {/* SWALLOW \n */}
+          <Tooltip
+            title="If enabled, new line characters will not be printed to the terminal display. If disabled, new line characters will be printed before any cursor movement occurs because of the new line, such that the new line character will be printed at the end of the existing line, not the start of the new line."
+            placement="top"
+            arrow
+          >
+            <FormControlLabel
+              control={
+                <Checkbox
+                  name="swallowNewLine"
+                  checked={app.settings.dataProcessing.swallowNewLine}
+                  onChange={(e) => {
+                    app.settings.dataProcessing.setSwallowNewLine(
+                      e.target.checked
+                    );
+                  }}
+                />
+              }
+              label="Swallow \n bytes"
+              sx={{ marginBottom: "10px" }}
+            />
+          </Tooltip>
+        </div>
+      </BorderedSection>
+
+      {/* =============================================================================== */}
+      {/* CARRIAGE RETURN SETTINGS */}
+      {/* =============================================================================== */}
+      <BorderedSection title="Carriage Returns">
+        <div
+          style={{
+            display: "flex",
+            flexDirection: "column",
+            maxWidth: "300px",
+            gap: "20px",
+          }}
+        >
+          {/* CARRIAGE RETURN CURSOR BEHAVIOR */}
+          <FormControl>
+            <FormLabel>When a \r byte is received:</FormLabel>
+            <RadioGroup
+              value={app.settings.dataProcessing.carriageReturnCursorBehavior}
+              onChange={(e) => {
+                app.settings.dataProcessing.setCarriageReturnBehavior(
+                  e.target.value as any
+                );
+              }}
+            >
+              {/* DO NOTHING */}
+              <Tooltip
+                title="Don't move the cursor at all when a carriage return character is received."
+                placement="right"
+                arrow
+              >
+                <FormControlLabel
+                  value={CarriageReturnCursorBehaviors.DO_NOTHING}
+                  control={<Radio />}
+                  label="Don't move the cursor"
+                />
+              </Tooltip>
+              {/* MOVE CURSOR TO START OF LINE */}
+              <Tooltip
+                title="Move the cursor to the start of the current line. A separate new line character is required if you want to move the cursor down one line."
+                placement="right"
+                arrow
+              >
+                <FormControlLabel
+                  value={CarriageReturnCursorBehaviors.CARRIAGE_RETURN}
+                  control={<Radio />}
+                  label="Move cursor to the start of the current line"
+                />
+              </Tooltip>
+              {/* CARRIAGE RETURN AND NEW LINE */}
+              <Tooltip
+                title="Move the cursor back to the start of the line and then down one line."
+                placement="right"
+                arrow
+              >
+                <FormControlLabel
+                  value={CarriageReturnCursorBehaviors.CARRIAGE_RETURN_AND_NEW_LINE}
+                  control={<Radio />}
+                  label="Move cursor to the start and then down one line."
+                />
+              </Tooltip>
+            </RadioGroup>
+          </FormControl>
+          {/* SWALLOW \r */}
+          <Tooltip
+            title="If enabled, carriage return characters will not be printed to the terminal display. If disabled, carriage return characters will be printed before any cursor movement occurs because of the carriage return, such that the carriage return character will be printed at the end of the row, not the start of the row."
+            placement="top"
+            arrow
+          >
+            <FormControlLabel
+              control={
+                <Checkbox
+                  name="swallowCarriageReturn"
+                  checked={app.settings.dataProcessing.swallowCarriageReturn}
+                  onChange={(e) => {
+                    app.settings.dataProcessing.setSwallowCarriageReturn(
+                      e.target.checked
+                    );
+                  }}
+                />
+              }
+              label="Swallow \r bytes"
+              sx={{ marginBottom: "10px" }}
+            />
+          </Tooltip>
+        </div>
+      </BorderedSection>
+
+      </div>
+
     </div>
   );
 }