+++ src/Util/Util.ts
@@ -10,3 +10,12 @@ export function generateRandomString(length: number) {
 
     return result;
 }
+
+export function stringToUint8Array(str: string) {
+    const arr = new Uint8Array(str.length);
+    for (let i = 0; i < str.length; i++) {
+        arr[i] = str.charCodeAt(i);
+    }
+
+    return arr;
+}